Update Import-XrmSolution.ps1
In some cases you want to first import solutions and only apply them later on using Start-XrmSolutionUpgrade. This is in the event of removal of entities in the newer version where other solutions are still depending on. 

When you import manually in power platform, you can choose for 'Upgrade' or 'Stage for upgrade'. Adding the parameter 'StageForUpgrade' to this script also gives you this option using scripting.